requirements.txt
playwright==1.47.0
python-dateutil
pytz
requests